slurm/interactive: Fix logic when user requests a specific node
When a user requests a specific node to run on without specifying
the number of CPUs, we need to make sure to use the default number
of CPUs so we can request an appropriate number of tasks. If a user
doesn't request CPUs explictly, they get 1, so let's just set it.