389/389_useradd.sh: Don't parse positional parameters in a function
Passing the arguments around seems to mess with the string of arguments,
for example if you pass in -f "Jean Pierre" the quoted parameter will be
two parameters by the time it reaches getopts! And we don't really even
need to have it as function because we're only using it once.