389/389_useradd.sh: Use $DEF_SHELL instead of $SHELL
$SHELL is essentially a reserved variable that is already present in the
environment, so we were effectively setting the shell of the the user to
be the same as *our* current shell. In effect this was always /bin/bash
but of course it's a major bug. I only noticed it because I was testing
it on macOS where my shell is /usr/local/bin/zsh!