ghetto-heal.sh: Check if file belongs using DM hash script
If the file exists and belongs on the current brick then we don't
really need to copy it to every other brick. Gluster will heal it
later on its own once the file is accessed.

I originally just copied all files from the brick to all other bri-
cks because I was lazy, and eventually I would use the rebalance.sh
script to remove files that didn't belong.